fix: 순환 참조 발생으로 인한 분리
